Reset guns after game over

+++ src/client.c
@@ -61,7 +61,7 @@ void client_disconnect(void)
 }
 
 void client_recv_updates(void (*on_update)(ClientUpdate), void (*on_remove)(int), 
-                         void (*on_kill)(int, int), void (*on_win)(int, int),
+                         void (*on_kill)(int, int), void (*on_win)(double, double),
                          void (*on_respawn)(int, double, double))
 {
     if (sock_fd == -1) return;
