+++ src/client.h
@@ -24,7 +24,7 @@ typedef struct {
 } ClientUpdate;
 
 void client_recv_updates(void (*on_update)(ClientUpdate), void (*on_remove)(int), void (*on_kill)(int, int)
-                                , void (*on_win)(int, int), void (*on_respawn)(int, double, double));
+                                , void (*on_win)(double, double), void (*on_respawn)(int, double, double));
 void client_recv_initial(Player *player, void (*on_update)(ClientUpdate));
 
 #endif
